start learn ansible

+++ inception/srcs/requirements/wordpress/tools/script.sh
@@ -1,31 +1,41 @@
 #!/bin/bash
-mkdir -p /var/www/html
-cd  /var/www/html/
+set -e
 
-wget https://wordpress.org/latest.tar.gz
+WWW_DIR=/var/www/html
+FLAG_FILE="$WWW_DIR/.wordpress_installed"
 
-tar xzvf latest.tar.gz
+mkdir -p "$WWW_DIR"
+cd "$WWW_DIR"
 
-mv -f wordpress/* .
+# Only download and extract WordPress if not already done
+if [ ! -f "$FLAG_FILE" ]; then
+    echo "Downloading and extracting WordPress..."
+    wget -q https://wordpress.org/latest.tar.gz
+    tar xzf latest.tar.gz
+    mv -f wordpress/* .
+    rm -rf latest.tar.gz wordpress
 
+    # Setup wp-config.php if it doesn't exist
+    if [ ! -f wp-config.php ]; then
+        cp wp-config-sample.php wp-config.php
 
-rm -rf latest.tar.gz wordpress
+        sed -i -r "s/database_name_here/$MARIADB_DATABASE/" wp-config.php
+        sed -i -r "s/username_here/$MARIADB_USER/" wp-config.php
+        sed -i -r "s/password_here/$MARIADB_PASSWORD/" wp-config.php
+        sed -i -r "s/localhost/mariadb/" wp-config.php
+    fi
 
-cp  wp-config-sample.php wp-config.php
+    # Download wp-cli only once
+    if ! command -v wp >/dev/null 2>&1; then
+        wget -q https://raw.githubusercontent.com/wp-cli/builds/gh-pages/phar/wp-cli.phar
+        chmod +x wp-cli.phar
+        mv wp-cli.phar /usr/local/bin/wp
+    fi
 
-
-sed -i -r "s/database_name_here/$MARIADB_DATABASE/1" wp-config.php
-sed -i -r "s/username_here/$MARIADB_USER/1" wp-config.php
-sed -i -r "s/password_here/$MARIADB_PASSWORD/1" wp-config.php
-sed -i -r "s/localhost/mariadb/1" wp-config.php
-
-
-
-#  configration wordpress
-
-wget https://raw.githubusercontent.com/wp-cli/builds/gh-pages/phar/wp-cli.phar
-chmod +x wp-cli.phar
-mv wp-cli.phar /usr/local/bin/wp
+    touch "$FLAG_FILE"
+else
+    echo "WordPress already installed, skipping setup."
+fi
 
 # Wait for MariaDB to be ready
 echo "Waiting for MariaDB to be ready..."
@@ -35,7 +45,7 @@ until mysqladmin ping -h"mariadb" --silent; do
 done
 echo "MariaDB is up!"
 
-# Configuration via wp-cli
+# Configure WordPress via wp-cli if not installed
 if ! wp core is-installed --allow-root; then
     echo "Installing WordPress..."
     wp core install --url="$DOMAIN_NAME" \
@@ -49,7 +59,7 @@ else
     echo "WordPress is already installed."
 fi
 
-# Only create user if it doesn't exist
+# Create additional user if it doesn't exist
 if ! wp user get "$WP_USER" --allow-root >/dev/null 2>&1; then
     echo "Creating additional WordPress user: $WP_USER"
     wp user create "$WP_USER" "$WP_EMAIL" --role=author --user_pass="$WP_PASS" --allow-root
@@ -57,9 +67,13 @@ else
     echo "User $WP_USER already exists."
 fi
 
-# change port in php 
-sed -i 's/listen = \/run\/php\/php7.3-fpm.sock/listen = 9000/g' /etc/php/7.3/fpm/pool.d/www.conf
-mkdir /run/php
+# Change PHP-FPM listen port if not already changed
+PHP_FPM_CONF="/etc/php/7.3/fpm/pool.d/www.conf"
+if ! grep -q "listen = 9000" "$PHP_FPM_CONF"; then
+    sed -i 's|listen = /run/php/php7.3-fpm.sock|listen = 9000|g' "$PHP_FPM_CONF"
+fi
 
-# run 
-/usr/sbin/php-fpm7.3 -F
+mkdir -p /run/php
+
+# Run PHP-FPM in foreground
+exec /usr/sbin/php-fpm7.3 -F
